Adds hooks to enable custom interpretation of specific fields.
By subclassing `DBFReader` and overriding `getFieldValue()` it is
possible to change the way a field is interpreted and returned to the
output.

This can solve some very weird corner cases.
An example is when for some reasons a character field is saved using an
encoding different from the one used in the rest of the file.